storage: Improve side panels on Details page
Now they use the same layout as the Devices, Drives, etc panels on the
Storage Overview page.
