+++ pkg/storaged/storage.css
@@ -399,24 +399,6 @@ td.vgroup-action {
     padding-left: 300px;
 }
 
-#detail-sidebar .table {
-    table-layout:fixed;
-}
-
-#detail-sidebar .table td {
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-}
-
-#detail-sidebar .table td img {
-    margin-top: 4px;
-}
-
-#detail-sidebar td.storage-action {
-    width: 51px;
-}
-
 .tab-actions {
     float: right;
     margin-top: -15px;
